fix(dock): contrast, memoization, defensive nil-check, and test organization
Third round of claude[bot] review on #2136:

1. text-slate-400 on the "Drag items here to add them" and "Items
   unavailable" popover messages violated CLAUDE.md's dark-surface text
   contrast rule (falls below WCAG AA 4.5:1 on slate-900). Changed both to
   text-slate-300.

2. `visibleItems` was computed as a plain const in the render body, so it
   was a new array reference every render — making it an always-changed
   useCallback dependency and defeating handleDragEnd's memoization. Wrapped
   in useMemo keyed on [folder.items, canAccessTool].

3. `shouldShowFolder`'s `items` parameter is typed as required, but
   Firestore/localStorage load dock data with a bare cast and no per-item
   validation — a legacy document could deliver `items: undefined` and crash
   the whole Dock render on `.some()`. Widened the param to accept
   `| undefined` and guard with `(items ?? [])`.

Also reorganizes test coverage per review feedback: `shouldShowFolder` and
`reorderPreservingHidden` (both pure helpers in folderPermissions.ts) had
tests scattered across FolderItem.test.tsx and Dock.test.tsx instead of a
dedicated co-located file. Consolidated into folderPermissions.test.ts,
matching the project's established extracted-module test-file convention.
