+++ components/layout/dock/FolderItem.test.tsx
@@ -4,7 +4,6 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { DndContext } from '@dnd-kit/core';
 import { SortableContext } from '@dnd-kit/sortable';
 import { FolderItem } from './FolderItem';
-import { reorderPreservingHidden } from './folderPermissions';
 import type { DockFolder } from '@/types';
 import { DEFAULT_GLOBAL_STYLE } from '@/types';
 
@@ -162,44 +161,5 @@ describe('FolderItem permission gating', () => {
 
     const tile = container.querySelector(`[data-folder-id="${folder.id}"]`);
     expect(tile?.querySelector('.lucide-folder-plus')).not.toBeInTheDocument();
-  });
-});
-
-describe('reorderPreservingHidden', () => {
-  // 'clock' and 'time-tool' stand in for visible items; 'weather' stands in
-  // for a permission-gated (hidden) item that must not move.
-  it('reorders visible items while leaving a hidden item at its original absolute index', () => {
-    // folder.items = ['clock', 'weather'(hidden), 'time-tool'];
-    // visibleItems = ['clock', 'time-tool']. Dragging 'time-tool' before
-    // 'clock' in visible-space must NOT shift 'weather' out of index 1 — a
-    // restored permission should find it exactly where it was left.
-    const result = reorderPreservingHidden(
-      ['clock', 'weather', 'time-tool'],
-      ['clock', 'time-tool'],
-      'time-tool',
-      'clock'
-    );
-
-    expect(result).toEqual(['time-tool', 'weather', 'clock']);
-  });
-
-  it('returns null when the dragged or drop-target type is not currently visible', () => {
-    const result = reorderPreservingHidden(
-      ['clock', 'weather', 'time-tool'],
-      ['clock', 'time-tool'],
-      'weather',
-      'clock'
-    );
-    expect(result).toBeNull();
-  });
-
-  it('reorders correctly when nothing is hidden (visibleItems === allItems)', () => {
-    const result = reorderPreservingHidden(
-      ['clock', 'time-tool'],
-      ['clock', 'time-tool'],
-      'clock',
-      'time-tool'
-    );
-    expect(result).toEqual(['time-tool', 'clock']);
   });
 });
